made changes to print_chessboard function

+++ 0x07-pointers_arrays_strings/7-print_chessboard.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "main.h"
 
-/**
+/** 
  * print_chessboard - a function that prints the chessboard
  * @a: array of chars
  *
@@ -10,7 +10,7 @@
 
 void print_chessboard(char (*a)[8])
 {
-	int x, y;
+	unsigned short int x, y;
 
 	for (x = 0; x < 8; x++)
 	{
